chore: separate fuzz tests
- own jest config, folder and script
- only run it in node v16 (since node v14 support is broken)
