+++ fuzz/regression.test/dom-parser.html.target.js/crash-doctype-in-element.xml
