+++ fuzz/regression.test/dom-parser.xml.target.js/crash-doctype-in-element.xml
