+++ jest.fuzz.config.js
@@ -88,19 +88,7 @@ module.exports = {
 	// preset: undefined,
 
 	// Run tests from one or more projects
-	// projects: [
-	// 	{
-	// 		displayName: 'test',
-	// 	},
-	// 	{
-	// 		displayName: {
-	// 			name: 'Jazzer.js',
-	// 			color: 'cyan',
-	// 		},
-	// 		testMatch: ['<rootDir>/test/fuzz/*.test.fuzz.js'],
-	// 		testRunner: '@jazzer.js/jest-runner',
-	// 	},
-	// ],
+	// projects: undefined,
 
 	// Use this configuration option to add custom reporters to Jest
 	// reporters: undefined,
@@ -150,10 +138,7 @@ module.exports = {
 	// testLocationInResults: false,
 
 	// The glob patterns Jest uses to detect test files
-	// testMatch: [
-	//   "**/__tests__/**/*.[jt]s?(x)",
-	//   "**/?(*.)+(spec|test).[tj]s?(x)"
-	// ],
+	testMatch: ['<rootDir>/fuzz/*.test.js'],
 
 	// An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
 	// testPathIgnorePatterns: [
@@ -161,13 +146,13 @@ module.exports = {
 	// ],
 
 	// The regexp pattern or array of patterns that Jest uses to detect test files
-	testRegex: ['^test', '.test.js$'],
+	// testRegex: ['test/**/*.test.js$'],
 
 	// This option allows the use of a custom results processor
 	// testResultsProcessor: undefined,
 
 	// This option allows use of a custom test runner
-	// testRunner: "jasmine2",
+	testRunner: '@jazzer.js/jest-runner',
 
 	// This option sets the URL for the jsdom environment. It is reflected in properties such as location.href
 	// testURL: "http://localhost",
