+++ package.json
@@ -33,6 +33,7 @@
 		"changelog": "auto-changelog --unreleased-only",
 		"start": "nodemon --watch package.json --watch lib --watch test --exec 'npm --silent run test && npm --silent run lint'",
 		"test": "jest",
+		"fuzz": "jest --config=./jest.fuzz.config.js",
 		"test:types": "cd examples/typescript-node-es6 && ./pretest.sh 3 && ./pretest.sh 4 && ./pretest.sh 5 && node dist/index.js",
 		"testrelease": "npm test && eslint lib",
 		"version": "./changelog-has-version.sh",
